refactor: address review feedback — remove try/except, non-fatal post-steps, move PSAP to parser
- Remove unnecessary try/except for create_fournos_resolve_entrypoint import
- Wrap _generate_psap_payload and _set_mlflow_metadata in try/except so
  post-benchmark failures don't mark the run as failed
- Move generate_psap_payload and write_psap_payload from psap_payload.py
  into parser.py, delete psap_payload.py
- Update README with postprocessing, MLflow, PSAP, Fournos, and presets docs

E2E tested on mehulvalidation (A100) and fournos-fg (L4, IBM VPC block).
