+++ projects/rhaiis/README.md
@@ -24,13 +24,23 @@ cleanup_isvc            Delete InferenceService + ServingRuntime
 On failure at any step, `capture_isvc_state` and `cleanup_isvc` still run
 (try/finally in the orchestration layer).
 
+After the benchmark, `run_and_postprocess()` automatically runs:
+- KPI computation (22 metrics via rhaiis postprocess plugin)
+- PSAP payload generation (model-furnace compatible JSON)
+- MLflow metadata tagging
+
 The benchmark step uses the canonical `projects.guidellm.toolbox.run_guidellm_benchmark`
 (shared with llm_d). rhaiis builds `guidellm_args` from its config and passes them
 to the canonical runner.
 
 ## Configuration
 
-- Project config: [`orchestration/config.yaml`](./orchestration/config.yaml)
+Config is split into `config.yaml` (base) and `config.d/` (per-domain):
+
+- [`orchestration/config.yaml`](./orchestration/config.yaml) — vaults, benchmarks, tests, caliper
+- [`orchestration/config.d/rhaiis.yaml`](./orchestration/config.d/rhaiis.yaml) — deploy defaults, vLLM args, images, gpu_types
+- [`orchestration/config.d/models.yaml`](./orchestration/config.d/models.yaml) — 52 model definitions
+- [`orchestration/config.d/workloads.yaml`](./orchestration/config.d/workloads.yaml) — workload profiles
 
 Key sections:
 
@@ -56,17 +66,56 @@ bin/run_ci rhaiis ci export-artifacts         # Caliper export to MLflow
 
 FournosJob example:
 ```yaml
+apiVersion: fournos.dev/v1
+kind: FournosJob
+metadata:
+  generateName: rhaiis-benchmark-
 spec:
+  owner: mehul
+  displayName: rhaiis-qwen3-benchmark
+  pipeline: forge-prepare-test
+  cluster: fournos-fg
+  priority: manual
+  hardware:
+    gpuType: l40s
+    gpuCount: 1
   executionEngine:
     forge:
       project: rhaiis
-      args: [test, llama-8b, profile1]
+      args: [nvidia, profile1]
       configOverrides:
-        rhaiis.images.nvidia: "quay.io/custom/image:tag"
+        tests.rhaiis.model_key: qwen3-0_6b
+        rhaiis.namespace: kserve-e2e-perf
+        rhaiis.deploy.image_pull_secret: npalaska-image-pull
+        rhaiis.deploy.storage_pvc: model-pvc
 ```
 
 Presets from `args` are applied via `project.args` → `presets.d/presets.yaml`.
-Config overrides (e.g. `rhaiis.images.nvidia`) are applied as variable overrides.
+Config overrides use dot-notation to set any nested config value.
+
+Available configOverrides:
+
+| Key | Description |
+|-----|-------------|
+| `tests.rhaiis.model_key` | Model key from config.d/models.yaml |
+| `rhaiis.namespace` | Kubernetes namespace |
+| `rhaiis.deploy.image_pull_secret` | Image pull secret name |
+| `rhaiis.deploy.storage_pvc` | PVC name for model storage |
+| `rhaiis.deploy.replicas` | Number of predictor replicas |
+| `rhaiis.images.nvidia` | vLLM container image override |
+| `rhaiis.vllm_args.tensor-parallel-size` | Tensor parallel size |
+| `benchmarks.guidellm.timeout` | Benchmark timeout in seconds |
+
+Monitoring Fournos jobs:
+```bash
+export KUBECONFIG=~/kubeconfigs/psap-automation-kubeconfig
+oc get fournosjobs -n psap-automation              # Job status
+oc get workloads -n psap-automation                # Queue status
+oc get pipelineruns -n psap-automation | grep <name>  # Pipeline progress
+oc logs -f <pod-name> -n psap-automation -c step-main # Live logs
+oc patch fournosjob <name> -n psap-automation \
+  --type merge -p '{"spec":{"shutdown":"Stop"}}'   # Stop a job
+```
 
 ## Main entrypoints
 
@@ -187,7 +236,7 @@ Artifacts are stored under `/tmp/forge_<timestamp>/002__run_guidellm_benchmark/a
 | Phi-4 | `phi-4`, `phi-4-fp8`, `-w4a16`, `-w8a8` | 1 |
 | Validation | `qwen3-0_6b` | 1 |
 
-Full list: `grep "^  [a-z]" orchestration/config.yaml`
+Full list: `grep "^[a-z]" orchestration/config.d/models.yaml`
 
 ## Workload profiles
 
@@ -199,3 +248,67 @@ From model_furnace `guidellm_profiles.iterations`:
 | `profile2` | 512 (stdev 128) | 2048 (stdev 512) | 1, 50, 100, 200, 300 | 450 |
 | `profile3` | 2048 | 128 | 1, 50, 100, 200, 300 | 450 |
 | `profile4` | 8000 | 1000 | 1, 25, 50, 75, 100 | 450 |
+
+## Postprocessing and KPIs
+
+After each benchmark run, the rhaiis postprocess plugin (`projects/rhaiis/postprocess/`)
+automatically computes 22 KPIs with model-furnace parity:
+
+| Category | KPIs |
+|----------|------|
+| Throughput | request_rate, tokens/s, input_tokens/s, output_tokens/s |
+| Latency | ttft (median, p95, p99), itl (median, p95, p99), tpot (median, p95, p99) |
+| Request latency | request_latency (median, p95) |
+| Token efficiency | input_tokens/request, output_tokens/request |
+| Request stats | concurrency, completed_requests, failed_requests |
+| Benchmark | duration, prompt_token_count_mean |
+
+KPIs are written to `kpis.jsonl` in the postprocess output directory.
+
+## PSAP payload
+
+A model-furnace compatible `PSAP_perf_<profile>_<accelerator>_<model>_<date>.json`
+is generated after each benchmark. It wraps the raw `benchmarks.json` with experiment
+metadata (model, vLLM image/args, accelerator, timestamps) for downstream compatibility.
+
+## MLflow export
+
+Results are exported to MLflow via the caliper pipeline. The export step runs as
+a separate CI command (`export-artifacts`) in the Fournos pipeline.
+
+MLflow tags set on each run:
+
+| Tag | Example |
+|-----|---------|
+| `project` | rhaiis |
+| `model_key` | qwen3-0_6b |
+| `hf_model_id` | Qwen/Qwen3-0.6B |
+| `accelerator` | nvidia |
+| `tensor_parallel_size` | 1 |
+| `vllm_image` | quay.io/aipcc/rhaiis/cuda-ubi9:3.4.0-ea.2 |
+| `vllm_version` | 3.4.0-ea.2-1773886296 |
+| `workload_key` | profile1 |
+| `rates` | 1,5,10 |
+| `guidellm_backend_type` | openai_http |
+
+MLflow experiment: `forge-rhaiis`
+
+## Presets
+
+Presets in `presets.d/presets.yaml` provide shortcuts for common configurations:
+
+```bash
+# Use presets instead of specifying model/workload/accelerator separately
+python3 -m projects.rhaiis.orchestration.cli test \
+  --preset llama-8b --preset profile1 \
+  --namespace kserve-e2e-perf
+
+# Available model presets: llama-8b, llama-70b, llama-405b, llama-4-scout,
+#   llama-4-maverick, granite-8b, mistral-24b, qwen25-7b, qwen3-235b,
+#   deepseek-r1, deepseek-v3, gpt-oss
+# Workload presets: profile1, profile2, profile3, profile4
+# Accelerator presets: nvidia, amd
+```
+
+Per-cluster presets (e.g. `presets.d/mehulvalidation.yaml`) set cluster-specific
+defaults like namespace, image pull secret, and service account.
